requirements.txt
streamlit
openai
supabase==2.3.1
cloudinary==1.37.0
python-dotenv==1.0.1